Fixed error with last commit

+++ Scratch.java
@@ -1976,7 +1976,7 @@ public class Scratch extends Actor
     {
         if (amount.intValue() < 0) {
             amount = 0;
-        } else if (amount > 100) {
+        } else if (amount.intValue() > 100) {
             amount = 100;
         }
         ghostEffect = amount.intValue();
